replaced clean script with Reset task. merged versions into dependencies file
