--- project/Versions.scala
@@ -1,15 +1,0 @@
-object Versions {
-  val Juju = "0.1.0-SNAPSHOT"
-  val Scala = "2.11.7"
-  val Akka = "2.4.1"
-  val ScalaLogging = "3.1.0"
-  val ScalaTest = "2.2.4"
-  val kafka = "0.8.2.1"
-  val ReactiveX = "0.25.0"
-  val KafkaPlugin = "0.4"
-  val AkkaKafkaPersistence = "2.3-14"
-  val LogBack = "1.1.3"
-  val Spray   = "1.3.3"
-  val SprayJson   = "1.3.2"
-  val ScalaReflect  = "2.11.7"
-}
