--- project/scripts/clean.sh
@@ -1,4 +1,0 @@
-echo 'deleting follow folders:'
-find . -name target -exec ls {} \;
-find . -name target -exec rm -rf {} \;
-rm -rf snapshots
